+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is simplified version of [Keep a Changelog](https://keepachangelog.co
 - Chapter progress now saved everything the page is changed
 - Adjust sorting order to be more consistent (@Astyyyyy)
 - Improve Local Source when loading from `android/data` (@lalalasupa0)
+- Refresh available extensions list when an extension repo is added or removed
 
 ### Fixes
 - Allow users to bypass onboarding's permission step if Shizuku is installed